pyproject.toml
[tool.poetry]
name = "scale-alibi"
version = "0.1.0"
description = ""
authors = ["Patrick Kage <patrick.r.kage@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "scale_alibi"}]

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.4.0"
torchvision = "^0.19.0"
pmtiles = "^3.3.0"
rich = "^13.7.1"
httpx = "^0.27.0"
asyncclick = "^8.1.7.2"
matplotlib = "^3.9.1"
pillow = "^10.4.0"
pydantic = "^2.8.2"
shapely = "^2.0.5"
rio-tiler = "^6.6.1"
rasterio = "^1.3.10"
geojson-pydantic = "^1.1.0"
supermercado = "^0.2.0"
einops = "^0.8.0"
wandb = "^0.17.5"
python-dotenv = "^1.0.1"
pystac-client = "^0.8.3"
pendulum = "^3.0.0"
boto3 = "^1.34.149"


[tool.poetry.scripts]
salibi = "scale_alibi:cli.cli"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
